Create mssage repository and test cases.
